Refactor NavBar to allow component-based sub-menus
Update Driver View to default to satellite view.
