+++ src/components/NavBar/AppNavBar.tsx
@@ -1,5 +1,5 @@
-import React, { useEffect, useState } from "react";
-import { NavLink as RouterNavLink, useNavigate } from "react-router-dom";
+import React, { useState } from "react";
+import { NavLink as RouterNavLink } from "react-router-dom";
 import fontawesome from '@fortawesome/fontawesome'
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faPowerOff, faUser } from '@fortawesome/fontawesome-free-solid'
@@ -21,12 +21,9 @@ import {
 import { useAuth0 } from "@auth0/auth0-react";
 import { NavBarBrand } from "./NavBarBrand";
 
-export const AppNavBar = () => {
-  const { getAccessTokenSilently } = useAuth0();
-  const navigate = useNavigate();
+export const AppNavBar = ({ additionalMenuItems }: { additionalMenuItems?: React.ReactNode; }) => {
 
   const [isOpen, setIsOpen] = useState(false);
-  const [token, setToken] = useState("");
 
   const {
     user,
@@ -38,82 +35,12 @@ export const AppNavBar = () => {
 
   const landingPageUri = process.env.REACT_APP_LANDING_PAGE_URL;
 
-  useEffect(() => {
-    if (token) {
-      return;
-    }
-
-    const audience = process.env.REACT_APP_AUTH0_AUDIENCE;
-    getAccessTokenSilently({
-      authorizationParams: {
-        audience: audience,
-      }
-    })
-      .then(async token => {
-        setToken(token);
-      });
-  }, [token, getAccessTokenSilently]);
-
   const logoutWithRedirect = () =>
     logout({
       logoutParams: {
         returnTo: landingPageUri,
       }
     });
-
-  const handleCreateCrissCross = async () => {
-    const url = `${process.env.REACT_APP_ROADRUNNER_REST_URL_BASE}/api/vehicle/create-crisscross`;
-    const body = {
-      degLatitude: 32.74666,
-      degLongitude: -97.319507,
-      kmRadius: 10.0,
-      vehicleCount: 15,
-    };
-
-    try {
-      const response = await fetch(url, {
-        method: 'post',
-        headers: {
-          "Content-Type": "application/json",
-          Authorization: `Bearer ${token}`,
-        },
-        body: JSON.stringify(body),
-      });
-
-      if (!response.ok) {
-        throw new Error(`HTTP error! Status: ${response.status}`);
-      }
-
-      const data = await response.json();
-      console.log("Success:", data);
-    } catch (error) {
-      console.error("Error creating criss-cross:", error);
-      alert("Failed to create criss-cross.");
-    }
-  };
-
-  const handleResetServer = async () => {
-    const url = `${process.env.REACT_APP_ROADRUNNER_REST_URL_BASE}/api/vehicle/reset-server`;
-    try {
-      const response = await fetch(url, {
-        method: 'get',
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      });
-
-      if (!response.ok) {
-        throw new Error(`HTTP error! Status: ${response.status}`);
-      }
-
-      const data = await response.json();
-      console.log("Success:", data);
-      navigate("/home");
-    } catch (error) {
-      console.error("Error resetting server:", error);
-      alert("Failed to reset server.");
-    }
-  };
 
   fontawesome.library.add(faPowerOff, faUser);
 
@@ -150,25 +77,7 @@ export const AppNavBar = () => {
             <Nav navbar>
               {isAuthenticated && (
                 <>
-                  <UncontrolledDropdown nav inNavbar>
-                    <DropdownToggle nav caret>
-                      Manage
-                    </DropdownToggle>
-                    <DropdownMenu>
-                      <DropdownItem
-                        id="crissCrossBtn"
-                        onClick={() => handleCreateCrissCross()}
-                      >
-                        Create criss-cross
-                      </DropdownItem>
-                      <DropdownItem
-                        id="resetServerBtn"
-                        onClick={() => handleResetServer()}
-                      >
-                        Reset server
-                      </DropdownItem>
-                    </DropdownMenu>
-                  </UncontrolledDropdown>
+                  {additionalMenuItems}
                   <UncontrolledDropdown nav inNavbar>
                     <DropdownToggle nav caret>
                       <img
